+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Test for quotes method raising OpenURI::HTTPError
 - SimpleCov for code coverage
 
 ### Changed